Add yoast token task and a fix for cookiehub
These changes fix some issues when installing kirjastokaista site.

Yoast token task has to be run after sync repo and before composer
install, so it more or less needs to be here.

Git refused to clone cookiehub plugin repo due to 'dubious ownership' of
base directory. Suggests adding this. A better alternative might be to
install the plugin with composer.